refactor: streamline Footer component by replacing static content with reusable subcomponents (FooterAbout, FooterLinks, FooterContact), enhance AppRoutes with clearer route organization and comments, simplify AuthContext by integrating useAuthMethods for authentication logic, and refactor Profile page to utilize AccountSettings component for improved structure and maintainability
