+++ src/components/layout/Footer.jsx
@@ -1,55 +1,22 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import { ROUTES } from '../../constants/routes';
+import FooterAbout from './footer/FooterAbout.jsx';
+import FooterLinks from './footer/FooterLinks.jsx';
+import FooterContact from './footer/FooterContact.jsx';
 
-const Footer = () => {
-  return (
-    <footer className="bg-gray-800 text-white">
-      <div className="container mx-auto px-4 py-8">
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Tea Haven</h3>
-            <p className="text-gray-300">
-              Discover the finest tea selections from around the world.
-            </p>
-          </div>
-          
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Quick Links</h3>
-            <ul className="space-y-2">
-              <li>
-                <Link to={ROUTES.HOME} className="text-gray-300 hover:text-white">
-                  Home
-                </Link>
-              </li>
-              <li>
-                <Link to={ROUTES.LOGIN} className="text-gray-300 hover:text-white">
-                  Sign In
-                </Link>
-              </li>
-              <li>
-                <Link to={ROUTES.REGISTER} className="text-gray-300 hover:text-white">
-                  Register
-                </Link>
-              </li>
-            </ul>
-          </div>
-          
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Contact</h3>
-            <p className="text-gray-300">
-              Email: support@teahaven.com<br />
-              Phone: (555) 123-4567
-            </p>
-          </div>
-        </div>
-        
-        <div className="border-t border-gray-700 mt-8 pt-8 text-center text-gray-300">
-          <p>&copy; {new Date().getFullYear()} Tea Haven. All rights reserved.</p>
-        </div>
+const Footer = () => (
+  <footer className="bg-gray-800 text-white">
+    <div className="container mx-auto px-4 py-8">
+      <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+        <FooterAbout />
+        <FooterLinks />
+        <FooterContact />
       </div>
-    </footer>
-  );
-};
+      
+      <div className="border-t border-gray-700 mt-8 pt-8 text-center text-gray-300">
+        <p>&copy; {new Date().getFullYear()} Tea Haven. All rights reserved.</p>
+      </div>
+    </div>
+  </footer>
+);
 
 export default Footer;
