+++ src/components/layout/footer/FooterAbout.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const FooterAbout = () => (
+  <div>
+    <h3 className="text-lg font-semibold mb-4">Tea Haven</h3>
+    <p className="text-gray-300">
+      Discover the finest tea selections from around the world.
+    </p>
+  </div>
+);
+
+export default FooterAbout;
