+++ src/components/layout/footer/FooterContact.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const FooterContact = () => (
+  <div>
+    <h3 className="text-lg font-semibold mb-4">Contact</h3>
+    <p className="text-gray-300">
+      Email: support@teahaven.com<br />
+      Phone: (555) 123-4567
+    </p>
+  </div>
+);
+
+export default FooterContact;
